backend/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Medical Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
SimpleITK>=2.2.1
scikit-image>=0.21.0

# MedSAM and Segment Anything
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
timm>=0.9.0

# Few-Shot Learning
# learn2learn removed due to installation issues

# Data Handling
pandas>=2.0.0
h5py>=3.9.0
albumentations>=1.3.1

# Metrics and Visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# API Server
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
